:twisted_rightwards_arrows: move classes into domain folders
