+++ src/Shipment/Http/Requests/ShipmentStatusCallbackRequest.php
@@ -7,17 +7,12 @@ namespace MyParcelCom\Integration\Http\Requests;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Support\Arr;
 use MyParcelCom\Integration\Exceptions\RequestInputException;
-use MyParcelCom\Integration\ShopId;
+use MyParcelCom\Integration\Shop\ShopId;
 use Ramsey\Uuid\Exception\InvalidUuidStringException;
 use Ramsey\Uuid\Uuid;
 
 class ShipmentStatusCallbackRequest extends FormRequest
 {
-    public function authorize(): bool
-    {
-        return true;
-    }
-
     /**
      * Get the validation rules that apply to the request.
      *
