+++ src/Shop/Http/Responses/ShopSetupResponse.php
